refactor foods delete into separate ajax call function and event listener function and in separate files
