+++ lib/FoodHandler.js
@@ -1,0 +1,13 @@
+const FoodCall = require('./FoodCall')
+
+
+const deleteHandler = () => {
+  $(document).on('click','.delete-food', function(){
+    let itemId = this.id
+    let deleteRow = this.parentElement.parentElement
+    FoodCall.deleteFood(itemId, deleteRow)
+  })
+}
+
+
+module.exports = { deleteHandler }
