+++ lib/foods.js
@@ -1,9 +1,10 @@
 const herokuURL = "https://quantified-self-node.herokuapp.com"
-const FoodTable = require('./FoodTable')
 const FoodCall  = require('./FoodCall')
+const FoodHandler  = require('./FoodHandler')
 
 $(document).ready(function(){
-  getFoods()
+  FoodCall.getFoods()
+  FoodHandler.deleteHandler()
   errorMessage()
   createFoodHandler()
   searchFoods()
@@ -11,17 +12,6 @@ $(document).ready(function(){
 });
 
 
-  function getFoods(){
-    $.get(`${herokuURL}/api/v1/foods`)
-      .then(function(foods){
-        foods.reverse()
-
-        FoodTable.buildTable()
-        FoodTable.addFoodData(foods)
-      })
-    }
-
-// function listenToEdit(){
 $(document).on('focus','#foodTable td', function(){
   $(this).data("initialText", $(this).html());
   $(document).on('blur','td', function() { // does this logic work if updating more than one td element at a time
@@ -47,26 +37,6 @@ $(document).on('focus','#foodTable td', function(){
     };
   });
 })
-
-// delete food handler
-let itemId;
-
-$(document).on('click','.delete-food', function(){
-  itemId = this.id
-  let deleteRow = this.parentElement.parentElement
-  console.log(`clicked delete for item with id: ${itemId}`);
-  $.ajax({
-    url: `${herokuURL}/api/v1/foods/${itemId}`,
-    type: 'DELETE',
-    success: function() {
-      deleteRow.remove(); // deletes foods not in meals
-    },
-    error: function() {
-      alert("Error")
-      deleteFoodMeal(itemId);
-    }
-  })
-});
 
 function searchFoods() {
 $('#searchInput').keyup(function() {
@@ -95,22 +65,22 @@ function errorMessage(){
 }
 
 // new food form handler
-function createFoodHandler(){
-  $(".add-food-form").submit(function( event ) {
-    if ( $( ".new-food-name" ).val().length === 0 & $( ".new-food-calories" ).val().length === 0) {
-      event.preventDefault();
-      $(".sub-name").show();
-      $(".sub-calories").show();
-    } else if ( $( ".new-food-name" ).val().length === 0) {
-      event.preventDefault();
-      $(".sub-calories").hide();
-      $(".sub-name").show();
-    } else if ( $( ".new-food-calories" ).val().length === 0) {
-      event.preventDefault();
-      $(".sub-name").hide();
-      $(".sub-calories").show();
-    } else {
-      FoodCall.postFood();
-    }
-  });
-}
+// function createFoodHandler(){
+//   $(".add-food-form").submit(function( event ) {
+//     if ( $( ".new-food-name" ).val().length === 0 & $( ".new-food-calories" ).val().length === 0) {
+//       event.preventDefault();
+//       $(".sub-name").show();
+//       $(".sub-calories").show();
+//     } else if ( $( ".new-food-name" ).val().length === 0) {
+//       event.preventDefault();
+//       $(".sub-calories").hide();
+//       $(".sub-name").show();
+//     } else if ( $( ".new-food-calories" ).val().length === 0) {
+//       event.preventDefault();
+//       $(".sub-name").hide();
+//       $(".sub-calories").show();
+//     } else {
+//       FoodCall.postFood();
+//     }
+//   });
+// }
