+++ lib/index.js
@@ -4,3 +4,4 @@ require('./index_styling.scss');
 require('./diary');
 require('./FoodTable');
 require('./FoodCall');
+require('./FoodHandler');
